Moved RangeMappedComparer to separate file.
